Remove debug flags from spec_utils so that every call to validate() does a validation, and add a "validation frequency" to timestep preprocessors.
Calls to validate() can be expensive. The debug flags in spec_utils allowed people to skip them, but rather than do that at the spec_utils library level, this CL moves that logic one level higher, to the users of that library. As part of that, the timestep preprocessors now govern how often they want to call validate().

PiperOrigin-RevId: 430059982
Change-Id: Iecd98e89a3f3c6ef1f4785fa261ca633f150a761