Enabled setting validation_frequency recursively on a tree of CompositeTimestepPreprocessors
This CL enables users to set/query validation_frequency on CompositeTimestepPreprocessors, as well as other TimestepPreprocessor containers e.g. CombineRewards.

The implementation depends on the new `as_list` method added to the TimestepPreprocessor API, which allows it to be traversed by the `tree` library.

E.g. you can now do
>>> tree.map_structure(
...   lambda p: p.set_validation_frequency(tsp.ValidationFrequency.NEVER),
...   processor.as_list())

PiperOrigin-RevId: 455362389
Change-Id: I0701d5c1a8dedc077e950458b44015201b8ef02f