Fix WeightedDiscretePoseDistribution and refactor to use canonical pose representation, and add test.
Previously this class threw the following error on valid inputs:
"ValueError: a must be 1-dimensional"
Due to attempting to invoke np.random.choice on 2-dimensional inputs.

We now correctly sample indices into the underlying pose array (with defensive copy).

Additionally this class previously operated on XYZ poseuler arrays rather than (pos, quat) tuples, which is inconsistent with the rest of the pose_distributions API, and prevented composition of classes as shown in the test.

PiperOrigin-RevId: 598849351
Change-Id: I898573730f6daae3875b1e7a367f554e1ac836e2